+++ comfit/src/com/test/mybatis/UserMyPageInfoController.java
@@ -16,6 +16,7 @@ public class UserMyPageInfoController
 	@Autowired
 	private SqlSession sqlSession;
 	
+	// 마이페이지 호출액션
 	@RequestMapping(value = "/user_mypage.action", method = RequestMethod.GET)
 	public String userMypage(Model model, HttpServletRequest request)
 	{
@@ -45,7 +46,7 @@ public class UserMyPageInfoController
 		return result;
 	}
 	
-	
+	// 마이페이지 내정보 수정 액션
 	@RequestMapping(value = "/user_information.action", method = RequestMethod.GET)
 	public String userInformation(Model model, HttpServletRequest request)
 	{
@@ -72,7 +73,7 @@ public class UserMyPageInfoController
 		
 	}
 	
-	// 구매내역
+	// 마이페이지 구매내역 액션
 	@RequestMapping(value = "/user_buylist.action", method = RequestMethod.GET)
 	public String userBuylist(Model model, HttpServletRequest request)
 	{
@@ -96,7 +97,7 @@ public class UserMyPageInfoController
 		return result;
 	}
 	
-	// 판매내역
+	// 마이페이지 판매내역 액션
 	@RequestMapping(value = "/user_selllist.action", method = RequestMethod.GET)
 	public String userSelllist(Model model, HttpServletRequest request)
 	{
@@ -121,7 +122,7 @@ public class UserMyPageInfoController
 		return result;
 	}
 	
-	// 입출금내역
+	// 마이페이지 입출금내역 액션
 	@RequestMapping(value = "/user_moneylist.action", method = RequestMethod.GET)
 	public String userMoneylist(Model model, HttpServletRequest request)
 	{
@@ -148,6 +149,7 @@ public class UserMyPageInfoController
 		return result;
 	}
 	
+	// 마이페이지 회원탈퇴 액션
 	@RequestMapping(value = "/user_drop.action", method = RequestMethod.GET)
 	public String userDrop(Model model)
 	{
@@ -158,6 +160,7 @@ public class UserMyPageInfoController
 		return result;
 	}
 	
+	// 마이페이지 찜한상품 액션
 	@RequestMapping(value = "/user_wish.action", method = RequestMethod.GET)
 	public String userWish(Model model)
 	{
@@ -168,7 +171,40 @@ public class UserMyPageInfoController
 		return result;
 	}
 	
+	// 마이페이지 입금 폼 액션
+	@RequestMapping(value = "/inputmoneyform.action", method = RequestMethod.GET)
+	public String inputMoney(Model model, HttpServletRequest request)
+	{
+		String result = null;
+		
+		try
+		{
+			HttpSession session = request.getSession();
+			String u_id = (String)session.getAttribute("u_id");
+			
+			IUserMyPage mypage = sqlSession.getMapper(IUserMyPage.class);
+			model.addAttribute("point", mypage.myPoint(u_id));
+			
+			result = "/WEB-INF/view/user/mypage/user_mypage_point_input.jsp";
+		}
+		catch (Exception e) 
+		{
+			System.out.println(e.toString());
+		}
+		
+		return result;
+	}
 	
+	// 마이페이지 출금 폼 액션
+	@RequestMapping(value = "/outputmoneyform.action", method = RequestMethod.GET)
+	public String outputMoney(Model model)
+	{
+		String result = null;
+		
+		result = "/WEB-INF/view/user/mypage/user_mypage_point_output.jsp";
+		
+		return result;
+	}
 	
 	
 }
