+++ comfit/src/com/test/mybatis/userDTO.java
@@ -6,21 +6,13 @@ public class userDTO
 	private String u_id, u_email, u_name, u_nickname, u_joindate
 	, u_tel, bancount, u_password, u_profile, u_address;
 
-	private String u_id, u_email, u_name, u_nickname, u_joindate, u_address
-	, u_tel, bancount, u_password, u_profile;
-
-
-
-	
 	// 차단회원 리스트 속성
 	private String ban_cate_name, ban_date, ban_end_date, rep_cate_name, ban_rea_detail;
 
 	// 이메일 인증 여부
 	private String auth , random;
 
-
 	
-
 
 
   public String getU_id()
@@ -121,16 +113,6 @@ public class userDTO
 	public void setU_profile(String u_profile)
 	{
 		this.u_profile = u_profile;
-	}
-
-	public String getU_address()
-	{
-		return u_address;
-	}
-
-	public void setU_address(String u_address)
-	{
-		this.u_address = u_address;
 	}
 
 	public String getBan_cate_name()
